refactor: extract transformer method
